requirements-dev.txt
cachetools==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-baton==1.5.6
django-bootstrap-form==3.4
django-easy-pdf==0.1.1
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.15.0
oauth2client==1.5.2
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyPDF2==1.26.0
python-decouple==3.3
pytz==2019.3
reportlab==3.5.32
rsa==4.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
uritemplate==3.0.1
webencodings==0.5.1
xhtml2pdf==0.2.3
